Seo title funtion and css tweak
Added funtion to use seo_title custom field as the page title on product
pages. Also uses the category description for the same purpose.

Although this works for the category pages a better solutuin
would be to add a separate field in the category edit page for page title.

Also added css files to adjust font in catagory header.