A few cosmetic changes & fixes
- Change number of cross-sell items on cart page to 2 items because 3 was
too much to fit in the given space.
- As a followup to above, I've removed the cross-sells on mobile view as it
displays above the cart totals area, forcing users to scroll past cross-sells
to view total cost... that is if they even think to scroll down. To some
people it looks like the page is broken
- Hid star ratings and original prices—on sale items—in the cross-sells area
- Fix problem with description in pinteres it share button
- Changed hamburger menu icon and interaction. Menu area now loads hidden on
smaller screens
- Changes sass config. Now compiles to main folder instead of subfolder. The
upshot is that style.css does not have to import from styles/screen.css
- Changed overall box-sizing to border-box.